+++ inc/classes/bs-database.php
@@ -4,11 +4,11 @@ class BS_Database {
     private $fields = array();
     public $connection;
     private $default_fields = array(
-        'bs-dbhost' => '',
-        'bs-dbname' => '',
-        'bs-dbprefix' => '',
-        'bs-dbuser' => '',
-        'bs-dbpass' => ''
+        'bs-dbhost' => BS_DBHOST,
+        'bs-dbname' => BS_DBNAME,
+        'bs-dbprefix' => BS_PREFIX,
+        'bs-dbuser' => BS_DBUSER,
+        'bs-dbpass' => BS_DBPASS
     );
 
     function __construct( $fields ) {
@@ -19,7 +19,7 @@ class BS_Database {
             $this->connection = new PDO( 'mysql:host='. $this->fields['bs-dbhost'] .';dbname='.$this->fields['bs-dbname'] , $this->fields['bs-dbuser'], $this->fields['bs-dbpass'] );
         }
         catch( Exception $e ) {
-            die( "MySQL fails : " );
+            exit( "<strong>MySQL fails :</strong> " . $e->getMessage() );
         }
     }
 
@@ -31,15 +31,46 @@ class BS_Database {
     /* Basic */
 
     function query( $req ) {
-        $res = $this->connection->query( $req );
-        return $res;
+        return $this->connection->query( $req );
+    }
+
+    function select( $query ) {
+        $res = $this->query( $query );
+        $rows = array();
+        foreach ( $res as $row ) {
+            $rows[$row];
+        }
+        return $row;
+    }
+
+    function insert( $database, $values = array() ) {}
+
+    function create_table( $table_name, $columns = array() ) {
+        // Set values if empty
+        if ( isset( $columns['id'] ) && empty( $columns['id'] ) ) {
+            $columns['id'] = 'int(11) unsigned NOT NULL AUTO_INCREMENT';
+        }
+        // Set default values
+        foreach ( $columns as $id => $value ) {
+            if ( empty( $value ) ) {
+                $columns[$id] = 'varchar(100) DEFAULT NULL';
+            }
+        }
+
+        // Build query
+        $query = "CREATE TABLE `".$this->fields['bs-dbprefix'].$table_name."` (";
+        foreach ( $columns as $id => $value ) {
+            $query .= "`".$id."` " . $value . ",";
+        }
+        $query .= "PRIMARY key( `id` ) ) CHARSET=utf8;";
+        return $this->query( $query );
     }
 
     /* Test */
     function test_fields( $fields ) {
         foreach ( $this->default_fields as $id => $field ) {
             if ( !isset( $fields[$id] ) ) {
-                $fields[$id] = '';
+                $fields[$id] = $field;
             }
         }
         return $fields;
@@ -47,7 +78,6 @@ class BS_Database {
 
     function test_install() {
         $test_table = $this->fields['bs-dbprefix'] . 'options';
-
         $db = $this->connection;
         $req = $db->prepare( "SHOW TABLES LIKE :db" );
         $req->execute( array( 'db' => $test_table ) );